Updated homepage organization. Photo links pretty
Significant updates to the picture links. The links now hold a constant
max width that scale down with the page to a minimum. This is not 100%
what I was looking for, I was hoping to make them stack. However, in
this context I think this is fine.

I have made no further updates to the recipes pages. I may play with
them more later, but for now I will consider them done. After all, the
rest is just text formatting, which I don't believe will be very inter-
esting without pictures. Being a project, I have no need to add stock
photography and I'm going to wait to see if this becomes part of the
rest of this project instead of jumping ahead right now.